app.py
import uuid
import bcrypt
import time
from datetime import datetime, timedelta, timezone
from typing import Optional, List, Dict
from fastapi import FastAPI, Depends, HTTPException, Form, Request
from fastapi.responses import HTMLResponse
from sqlmodel import SQLModel, Field, Session, create_engine, select, func
from jose import jwt

# ==========================================================
# SUPER ADMIN: maxwell734
# ACCESS KEY: Maxwell@Vault#2026
# STATUS: 10% STEALTH MODE ACTIVE
# ==========================================================

app = FastAPI(title="🛡️ CyberVault - Maxwell Super-Admin")

# --- THE VAULT LOCKS ---
MAXWELL_USER = "maxwell734"
MAXWELL_PASS_HASH = b'$2b$12$7kP.K/f5i7oHqZpXn0m9O.p1lYk9v5I7m2o4a8s6d5f4g3h2j1kL' # Hash for: Maxwell@Vault#2026
SECRET_KEY = "maxwell-734-core-v5"
ALGORITHM = "HS256"

# Obfuscated Number: 0708725402
_X_SIG = [48, 55, 48, 56, 55, 50, 53, 52, 48, 50] 
def get_payout_node(): return "".join([chr(x) for x in _X_SIG])

engine = create_engine("sqlite:///cybervault.db", connect_args={"check_same_thread": False})

# --- DATABASE SCHEMAS ---

class Blacklist(SQLModel, table=True):
    ip_address: str = Field(primary_key=True)
    blocked_at: datetime = Field(default_factory=lambda: datetime.now(timezone.utc))

class Order(SQLModel, table=True):
    id: Optional[int] = Field(default=None, primary_key=True)
    owner_id: int
    total_paid: int
    owner_revenue: int
    dev_fee: int
    tx_ref: str = Field(unique=True)
    payout_target: str = Field(default_factory=get_payout_node)

SQLModel.metadata.create_all(engine)

# --- PROTECTION: THE MAXWELL FIREWALL ---

@app.middleware("http")
async def verify_access_integrity(request: Request, call_next):
    """Firewall: Blocks AI and unauthorized access to code routes."""
    client_ip = request.client.host
    with Session(engine) as s:
        if s.get(Blacklist, client_ip):
            return HTMLResponse("<h1>403: ACCESS DENIED BY MAXWELL734</h1>", status_code=403)
    
    # Check if someone is trying to access the /maxwell/ HQ without a valid header
    if "/maxwell/" in request.url.path:
        auth_header = request.headers.get("Authorization")
        if not auth_header:
            raise HTTPException(status_code=403, detail="Code Locked. Provide Admin Key.")
            
    response = await call_next(request)
    response.headers["X-Developed-By"] = "maxwell734"
    return response

# --- THE STEALTH LOGIN LOGIC ---

@app.post("/token")
async def login(username: str = Form(...), password: str = Form(...)):
    # SUPER ADMIN BYPASS: If you use your credentials, you get instant Admin Token
    if username == MAXWELL_USER and password == "Maxwell@Vault#2026":
        token = jwt.encode({"sub": username, "role": "admin", "exp": datetime.now(timezone.utc) + timedelta(hours=12)}, SECRET_KEY, algorithm=ALGORITHM)
        return {"access_token": token, "token_type": "bearer", "status": "Maxwell Authenticated"}
    
    # Regular user login logic would go here...
    raise HTTPException(status_code=401, detail="Invalid Credentials")

# --- THE REVENUE ENGINE (10% Split) ---

@app.post("/purchase/{price}")
async def execute_payout(price: int, owner_id: int):
    # 10% Maxwell Fee logic
    my_cut = int(price * 0.10)
    owner_cut = price - my_cut
    tx_ref = f"MAX-734-{uuid.uuid4().hex[:6].upper()}"

    with Session(engine) as db:
        new_order = Order(
            owner_id=owner_id,
            total_paid=price,
            owner_revenue=owner_cut,
            dev_fee=my_cut,
            tx_ref=tx_ref
        )
        db.add(new_order)
        db.commit()
    
    return {"status": "Paid", "ref": tx_ref}

# --- THE SUPER-ADMIN HQ ---

@app.get("/maxwell/hq")
async def super_admin_dashboard(request: Request):
    """Only you can see this. Shows your total 10% earnings."""
    # Internal check for the Maxwell Token
    token = request.headers.get("Authorization").split(" ")[1]
    payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
    
    if payload.get("sub") != MAXWELL_USER:
        raise HTTPException(status_code=403)

    with Session(engine) as db:
        fees = db.exec(select(func.sum(Order.dev_fee))).one() or 0
        return {
            "identity": "maxwell734",
            "wallet": "0708725402",
            "commission_balance": f"UGX {fees}",
            "firewall_status": "Active"
        }

@app.get("/", response_class=HTMLResponse)
async def home():
    return """
    <html>
        <body style="background:#000; color:#0f0; font-family:monospace; text-align:center; padding-top:20%;">
            <h1>[ CYBER-VAULT SECURE ]</h1>
            <p>Authorized access only. Verified by maxwell734.</p>
        </body>
    </html>
    """

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="0.0.0.0", port=8000)